+++ installers/vs_install_ruby.sh
@@ -6,15 +6,8 @@
 ### $License: Public Domain $
 ###
 
-_cmd() {
-    echo '$' $1
-    if eval $1; then
-        return 0
-    else
-        echo "** FAILED: $1" 2>&1
-        return 1
-    fi
-}
+. $HOME/.vs/installers/vs_install.sh
+
 
 _install_ruby() {
     ## arguments and variables
@@ -58,8 +51,12 @@ _install_ruby() {
         ;;
     esac
     local base="ruby-$version"
-    local url="ftp://ftp.ruby-lang.org/pub/ruby/$ver/$base.$ext"
-    _cmd "wget -N $url"                           || return 1
+    local filename="$base.$ext"
+    local url="ftp://ftp.ruby-lang.org/pub/ruby/$ver/$filename"
+    if [ ! -f "$filename" ]; then
+        local down=`__vs_downloader "-ORL" "-N"`  || return 1
+        _cmd "$down $url"                         || return 1
+    fi
     _cmd "rm -rf $base"                           || return 1
     _cmd "$tar $base.$ext"                        || return 1
     _cmd "cd $base/"                              || return 1
@@ -87,7 +84,8 @@ _install_ruby() {
         y*|Y*)
             base="rubygems-1.8.11"
             url="http://production.cf.rubygems.org/rubygems/$base.tgz"
-            _cmd "wget -N $url"                              || return 1
+            local down=`__vs_downloader "-ORL" "-N"`         || return 1
+            _cmd "$down $url"                                || return 1
             _cmd "tar xzf $base.tgz"                         || return 1
             _cmd "cd $base/"                                 || return 1
             _cmd "$prefix/bin/ruby setup.rb"                 || return 1
